Refactor test setup to use testingLibrarySetup.js
- Removed src/setupTests.js and consolidated test setup into src/testUtils/vitest/testingLibrarySetup.js.
- Updated imports in test files and vite.config.js to reference the new setup file.

This change centralises and streamlines test environment configuration, and removes confusion with similar file names.
