+++ src/testUtils/vitest/testingLibrarySetup.js
@@ -1,5 +1,11 @@
+import { configure } from '@testing-library/react'
 import { vi } from 'vitest'
+import '@testing-library/jest-dom'
 
+// Configure testing library
+configure({ testIdAttribute: 'data-cy' })
+
+// Setup mocks
 export const setupMocks = () => {
 	global.fetch = vi.fn()
 
@@ -13,3 +19,6 @@ export const setupMocks = () => {
 		writable: true,
 	})
 }
+
+// Auto-run mocks setup
+setupMocks()
