+++ vite.config.js
@@ -83,7 +83,7 @@ export default defineConfig({
 	test: {
 		globals: true,
 		environment: 'jsdom',
-		setupFiles: './src/setupTests.js',
+		setupFiles: './src/testUtils/vitest/testingLibrarySetup.js',
 		coverage: {
 			provider: 'v8',
 			reportsDirectory: './coverage',
